+++ shop/cart/cart_modifiers_base.py
@@ -113,15 +113,20 @@ class BaseCartModifier(object):
         a tuple. The decimal should be the amount that should get added to the
         current subtotal. It can be a negative value.
 
+        An optional third tuple element can be used to store extra data of any
+        kind, which must be serializable as JSON.
+
         In case your modifier is based on the current price (for example in
         order to compute value added tax for this cart item only) your
         override can access that price via ``cart_item.current_total``.
 
         A tax modifier would do something like this:
-        >>> return ('taxes', Decimal(9))
+        >>> return ('taxes', Decimal(9), {'rate': Decimal(10), 'identifier': 'V.A.T.'})
+        Note that the third element in this tuple is optional.
 
         And a rebate modifier would do something along the lines of:
-        >>> return ('rebate', Decimal(-9))
+        >>> return ('rebate', Decimal(-9), {'rate': Decimal(3), 'identifier': 'Discount'})
+        Note that the third element in this tuple is optional.
 
         More examples can be found in shop.cart.modifiers.*
         """